+++ srcs/get_input/seek_n_replace.c
@@ -28,7 +28,7 @@ int		seek_n_repl_str(t_edit *e, size_t x, char *word, size_t i)
 	hist = hist->next;
 	if (!(hist = search_by_occurence(hist, word)))
 		return (FAILURE);
-	if (!(tmp = ft_strpull(e->hist->s, e->hist->s + x, i - x - 1, hist->s)))
+	if (!(tmp = ft_strpull(e->hist->s, e->hist->s + x, (int)(i - x - 1), hist->s)))
 		return (FAILURE);
 	ft_strdel(&e->hist->s);
 	e->hist->s = tmp;
@@ -54,7 +54,7 @@ int		seek_n_repl_nb(t_edit *e, size_t x, int nb, size_t size)
 		hist = search_by_number_from_last(hist, nb);
 	if (!hist)
 		return (FAILURE);
-	if (!(tmp = ft_strpull(e->hist->s, e->hist->s + x, size, hist->s)))
+	if (!(tmp = ft_strpull(e->hist->s, e->hist->s + x, (int)size, hist->s)))
 		return (error("Memory allocation failed", NULL));
 	ft_strdel(&e->hist->s);
 	e->hist->s = tmp;
